+++ core/ray.h
@@ -11,10 +11,10 @@ namespace core {
 class Ray {
 public:
   Ray(const Vec3 &origin = Vec3(0, 0, 0), const Vec3 &direction = Vec3(0, 0, 1),
-      double energy = 0);
+      float energy = 0);
 
-  Vec3 at(double time) const;
-  double phaseAt(double freq, double time) const;
+  Vec3 at(float time) const;
+  float phaseAt(float freq, float time) const;
 
   friend std::ostream &operator<<(std::ostream &os, const Ray &srcRay);
   bool operator==(const Ray &other) const;
@@ -23,17 +23,17 @@ public:
   Vec3 origin() const;
   void setDirection(const Vec3 &direction);
   Vec3 direction() const;
-  void setEnergy(double num);
-  double energy() const;
+  void setEnergy(float num);
+  float energy() const;
 
 private:
   Vec3 origin_, direction_;
-  double energy_;
+  float energy_;
 };
 
 struct RayHitData {
   RayHitData() = default;
-  RayHitData(double t, const Vec3 &norm, const Ray &ray, double freq)
+  RayHitData(float t, const Vec3 &norm, const Ray &ray, float freq)
       : time(t), ray_(ray), normal_(norm), frequency(freq){};
   ~RayHitData() = default;
   RayHitData(const RayHitData &) = default;
@@ -45,9 +45,9 @@ struct RayHitData {
   Vec3 collisionPoint() const { return ray_.at(time); }
   Vec3 direction() const { return ray_.direction(); }
   Vec3 origin() const { return ray_.origin(); }
-  double energy() const { return ray_.energy(); }
-  double phase() const { return ray_.phaseAt(frequency, time); }
-  double frequency, time;
+  float energy() const { return ray_.energy(); }
+  float phase() const { return ray_.phaseAt(frequency, time); }
+  float frequency, time;
 
 private:
   Vec3 normal_;
