+++ core/vec3.cpp
@@ -5,7 +5,7 @@
 
 namespace core {
 
-Vec3::Vec3(std::initializer_list<double> initList) {
+Vec3::Vec3(std::initializer_list<float> initList) {
   if (initList.size() != 3) {
     std::stringstream ss1;
     ss1 << "Vec3 could not ve constructed: initializer_list size: "
@@ -24,16 +24,16 @@ Vec3 operator+(const Vec3 &left, const Vec3 &right) {
   return Vec3(left.x() + right.x(), left.y() + right.y(), left.z() + right.z());
 }
 
-Vec3 operator+(const Vec3 &left, double right) {
+Vec3 operator+(const Vec3 &left, float right) {
   return Vec3(left.x() + right, left.y() + right, left.z() + right);
 }
 
-Vec3 operator+(double left, const Vec3 &right) { return right + left; }
+Vec3 operator+(float left, const Vec3 &right) { return right + left; }
 
 Vec3 operator-(const Vec3 &left, const Vec3 &right) {
   return Vec3(left.x() - right.x(), left.y() - right.y(), left.z() - right.z());
 }
-Vec3 operator-(const Vec3 &left, double num) {
+Vec3 operator-(const Vec3 &left, float num) {
   return Vec3(left.x() - num, left.y() - num, left.z() - num);
 }
 
@@ -44,7 +44,7 @@ Vec3 &Vec3::operator+=(const Vec3 &other) {
   return *this;
 }
 
-Vec3 &Vec3::operator+=(double num) {
+Vec3 &Vec3::operator+=(float num) {
   x_ += num;
   y_ += num;
   z_ += num;
@@ -58,29 +58,29 @@ Vec3 &Vec3::operator-=(const Vec3 &other) {
   return *this;
 }
 
-Vec3 &Vec3::operator-=(double num) {
+Vec3 &Vec3::operator-=(float num) {
   x_ -= num;
   y_ -= num;
   z_ -= num;
   return *this;
 }
 
-Vec3 operator*(double num, const Vec3 &vec) {
+Vec3 operator*(float num, const Vec3 &vec) {
   return Vec3(vec.x() * num, vec.y() * num, vec.z() * num);
 }
 
-Vec3 operator*(const Vec3 &vec, double num) {
+Vec3 operator*(const Vec3 &vec, float num) {
   return Vec3(vec.x() * num, vec.y() * num, vec.z() * num);
 }
 
-Vec3 &Vec3::operator*=(double num) {
+Vec3 &Vec3::operator*=(float num) {
   x_ *= num;
   y_ *= num;
   z_ *= num;
   return *this;
 }
 
-Vec3 operator/(const Vec3 &vec, double num) {
+Vec3 operator/(const Vec3 &vec, float num) {
   if (num <= constants::kAccuracy) {
     std::stringstream ss;
     ss << "Divider of the Vec3 object can't be close or equal to 0. Object: "
@@ -91,15 +91,15 @@ Vec3 operator/(const Vec3 &vec, double num) {
   }
 }
 
-double Vec3::scalarProduct(const Vec3 &other) const {
+float Vec3::scalarProduct(const Vec3 &other) const {
   return x_ * other.x() + y_ * other.y() + z_ * other.z();
 }
 Vec3 Vec3::crossProduct(const Vec3 &other) const {
   return Vec3(y_ * other.z() - z_ * other.y(), z_ * other.x() - x_ * other.z(),
               x_ * other.y() - y_ * other.x());
 }
-double Vec3::magnitudeSquared() const { return x_ * x_ + y_ * y_ + z_ * z_; }
-double Vec3::magnitude() const { return std::sqrt(this->magnitudeSquared()); }
+float Vec3::magnitudeSquared() const { return x_ * x_ + y_ * y_ + z_ * z_; }
+float Vec3::magnitude() const { return std::sqrt(this->magnitudeSquared()); }
 Vec3 Vec3::normalize() const { return (*this) / magnitude(); }
 
 bool Vec3::operator==(const Vec3 &other) const {
@@ -121,11 +121,11 @@ Vec3 &Vec3::operator=(const Vec3 &other) {
   return *this;
 }
 
-void Vec3::setX(double num) { x_ = num; }
-double Vec3::x() const { return x_; }
-void Vec3::setY(double num) { y_ = num; }
-double Vec3::y() const { return y_; }
-void Vec3::setZ(double num) { z_ = num; }
-double Vec3::z() const { return z_; }
+void Vec3::setX(float num) { x_ = num; }
+float Vec3::x() const { return x_; }
+void Vec3::setY(float num) { y_ = num; }
+float Vec3::y() const { return y_; }
+void Vec3::setZ(float num) { z_ = num; }
+float Vec3::z() const { return z_; }
 
 } // namespace core
